# project/backend/document/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "个人贷款管理系统 API",
    "description": "个人贷款管理系统后端API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/users/refresh-token": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "刷新token",
        "description": "刷新用户的访问令牌",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "description": "用户ID",
                "format": "int64"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/blacklist/remove": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "解除黑名单",
        "description": "管理员将用户从黑名单中解除",
        "operationId": "removeFromBlackList",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "用户ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/blacklist/add": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "加入黑名单",
        "description": "管理员将用户加入黑名单",
        "operationId": "addToBlackList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToBlackListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/avatar": {
      "post": {
        "tags": [
          "用户管理"
        ],
        "summary": "上传头像",
        "description": "用户上传头像图片",
        "operationId": "uploadAvatar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "头像文件",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/me": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取当前用户信息",
        "description": "用户查看自己的详细信息",
        "operationId": "getCurrentUserInfo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUserSelfResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "用户管理"
        ],
        "summary": "更新用户信息",
        "description": "用户更新自己的信息（仅限用户名和头像）",
        "operationId": "updateCurrentUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUserSelfResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}/detail": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "获取用户详情",
        "description": "管理员获取指定用户的详细信息",
        "operationId": "getAdminUserDetail",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "用户ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultUserDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/search-by-credit": {
      "get": {
        "tags": [
          "用户管理"
        ],
        "summary": "按信用分搜索用户",
        "description": "按信用分表达式搜索用户",
        "operationId": "searchByCreditScore",
        "parameters": [
          {
            "name": "expr",
            "in": "query",
            "description": "信用分表达式",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSearchDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200,
            "refType": null
          },
          "data": {
            "type": "string",
            "description": "响应数据",
            "refType": null
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功",
            "refType": null
          }
        },
        "description": "API响应结果"
      },
      "AddToBlackListRequest": {
        "required": [
          "blackLevel",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64",
            "example": 1,
            "refType": null
          },
          "blackLevel": {
            "type": "integer",
            "description": "黑名单等级（1: 轻度, 2: 中度, 3: 重度）",
            "format": "int32",
            "example": 2,
            "enum": [
              1,
              2,
              3
            ],
            "refType": null
          }
        },
        "description": "加入黑名单请求"
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "正常",
              "已逾期",
              "已完成"
            ],
            "refType": null
          },
          "repaidAmount": {
            "type": "number",
            "refType": null
          },
          "loanAmount": {
            "type": "number",
            "refType": null
          },
          "interestRate": {
            "type": "number",
            "refType": null
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ],
            "refType": null
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "term": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "currentTerm": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "contract": {
            "type": "string",
            "refType": null
          },
          "overdueDays": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
...
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "refType": null
          },
          "avatar": {
            "type": "string",
            "refType": null
          }
        },
        "description": "用户更新信息"
      },
      "ApiResultUserSelfResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/UserSelfResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "UserSelfResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userName": {
            "type": "string",
            "refType": null
          },
          "avatar": {
            "type": "string",
            "refType": null
          }
        },
        "description": "响应数据"
      },
      "ApiResultUserDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/UserDetailResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "LoanApplication": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "审核中",
              "已通过",
              "AI拒绝",
              "人工拒绝",
              "已取消"
            ],
            "refType": null
          },
          "loanAmount": {
            "type": "number",
            "refType": null
          },
          "interestRate": {
            "type": "number",
            "refType": null
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "term": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ],
            "refType": null
          },
          "rejectReason": {
            "type": "string",
            "refType": null
          },
          "applyTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "reviewTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
      "User": {
        "required": [
          "password",
          "phone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "userName": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "refType": null
          },
          "avatar": {
            "type": "string",
            "refType": null
          },
          "phone": {
            "pattern": "^1[3-9]\\d{9}$",
            "type": "string",
            "refType": null
          },
          "password": {
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[!@#$%&*()_+=])[A-Za-z\\d!@#$%&*()_+=]{8,20}$",
            "type": "string",
            "refType": null
          },
          "role": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
      "UserCert": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "idCard": {
            "type": "string",
            "refType": null
          },
          "creditScore": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "bankCardId": {
            "type": "string",
            "refType": null
          },
          "workCertId": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "triCertId": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "immovableCertId": {
            "type": "integer",
            "format": "int32",
            "refType": null
          }
        }
      },
      "UserDetailResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User",
            "refType": "User"
          },
          "userCert": {
            "$ref": "#/components/schemas/UserCert",
            "refType": "UserCert"
          },
          "loanApplication": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanApplication"
            },
            "refType": "LoanApplication"
          },
          "order": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            },
            "refType": "Order"
          }
        },
        "description": "响应数据"
      },
...
      "UserSearchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "creditScore": {
            "type": "integer",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}